Translate the original WOZMON from 6502 to gbz80 assembly
This is translated almost line for line making heavy use of the hl
register. It is working and tested on my gameboy virtual terminal
(https://github.com/RubenZwietering/gameboy-virtual-terminal). If you
want to use it in your own project all you have to do is create a CHRIN
and CHROUT subroutine.