Use register e for variable stl and d for sth
It seems it is unnecessary to use seperate variables for XAM and STOR.
Perhaps this could be changed in the original as well for a small
optimazation.
Sidenote: the binary size for this code is now 232 bytes!
Smaller than the original 6502 wozmon even, however this does not have
chrin and chrout subroutines of course.